Update to removing cart item to remove based on count specified.

+++ database.py
@@ -227,15 +227,27 @@ class Database:
           self.connection.execute(ins)
           return 1
   
-  def remove_item_from_cart(self, uid, iid):
-    if uid > 0 and iid > 0:
+  def remove_item_from_cart(self, uid, iid, count):
+    if uid > 0 and iid > 0 and count > 0:
       statement = text("SELECT * FROM users WHERE users.id = '{}'".format(uid))
       res = self.connection.execute(statement).fetchone()
       if res:
-        statement = text("SELECT * FROM store_item WHERE store_item.id = '{}'".format(iid))
+        cart_id = self.get_user_cart(uid).id
+        statement = text("SELECT * FROM shopping_cart_item WHERE item_id = '{}'".format(iid))
         res = self.connection.execute(statement).fetchone()
-        if res:
-          cart_id = self.get_user_cart(uid).id
+        print(res.count)
+        print(count)
+        if res.count <= count:
           statement = text("DELETE FROM shopping_cart_item WHERE item_id = '{}'".format(iid))
-          self.connection.execute(statement)
-          return 1
+          res = self.connection.execute(statement)
+          if res:
+            return 1
+          else:
+            return -1
+        else:
+          statement = text("UPDATE shopping_cart_item SET count = count - '{}' WHERE item_id = '{}'".format(count, iid))
+          res = self.connection.execute(statement)
+          if res:
+            return 1
+          else:
+            return -1
